Add message account id to receiver data and move to message queries
Remove redundant bulletin implementation
