+++ service/src/main/kotlin/fi/espoo/evaka/Audit.kt
@@ -129,13 +129,13 @@ enum class Audit(
     MessagingBulletinDraftDelete("messaging.bulletin.draft.delete"),
     MessagingBulletinSend("messaging.bulletin.send"),
     MessagingBulletinRead("messaging.bulletin.read"),
-    MessagingBulletinReceiversRead("messaging.bulletin.receivers.read"),
     MessagingBulletinMarkRead("messaging.bulletin.mark-read"),
     MessagingUnreadCountRead("messaging.bulletin.unread-count.read"),
     MessagingSenderOptionsRead("messaging.bulletin.sender-options.read"),
     MessagingMyAccountsRead("messaging.messages.my-accounts.read"),
     MessagingUnreadMessagesRead("messaging.messages.unread.read"),
     MessagingMarkMessagesReadWrite("messaging.messages.mark-read.write"),
+    MessagingMessageReceiversRead("messaging.message.receivers.read"),
     MessagingReceivedMessagesRead("messaging.messages.received-messages.read"),
     MessagingSentMessagesRead("messaging.messages.sent-messages.read"),
     MessagingNewMessageWrite("messaging.messages.new-message.write"),
